+++ apps/java/libraries/case-engine/src/main/java/com/wks/caseengine/jpa/entity/CaseDefinitionEntity.java
@@ -4,9 +4,11 @@ import java.util.List;
 import java.util.UUID;
 
 import com.wks.caseengine.cases.definition.CaseStage;
+import com.wks.caseengine.cases.definition.DocumentRequirement;
 import com.wks.caseengine.event.ActionHook;
 import com.wks.caseengine.jpa.entity.converter.ActionHookListConverter;
 import com.wks.caseengine.jpa.entity.converter.CaseStageListConverter;
+import com.wks.caseengine.jpa.entity.converter.DocumentRequirementListConverter;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Convert;
@@ -50,5 +52,9 @@ public class CaseDefinitionEntity {
 	@Column(name="case_hooks", columnDefinition = "TEXT")
 	@Convert(converter = ActionHookListConverter.class)
 	private List<ActionHook> caseHooks;
+
+	@Column(name="required_documents", columnDefinition = "TEXT")
+	@Convert(converter = DocumentRequirementListConverter.class)
+	private List<DocumentRequirement> requiredDocuments;
 
 }