+++ apps/java/libraries/case-engine/src/main/java/com/wks/caseengine/jpa/entity/converter/CaseDefinitionConverter.java
@@ -16,6 +16,7 @@ public class CaseDefinitionConverter {
         caseDefinition.setDeployed(entity.getDeployed());
         caseDefinition.setStages(entity.getStages());
         caseDefinition.setCaseHooks(entity.getCaseHooks());
+        caseDefinition.setRequiredDocuments(entity.getRequiredDocuments());
 
         return caseDefinition;
     }
@@ -31,6 +32,7 @@ public class CaseDefinitionConverter {
         entity.setDeployed(caseDefinition.getDeployed());
         entity.setStages(caseDefinition.getStages());
         entity.setCaseHooks(caseDefinition.getCaseHooks());
+        entity.setRequiredDocuments(caseDefinition.getRequiredDocuments());
 
         return entity;
     }